PARTE_4/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.1
Flask==2.0.3
Flask-PyMongo==2.3.0
idna==3.3
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.1
PyJWT==2.3.0
pymongo==4.0.2
python-dotenv==0.19.2
requests==2.27.1
urllib3==1.26.9
Werkzeug==2.0.3
